requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.1
flask-mongoengine @ git+https://github.com/idoshr/flask-mongoengine.git@e244408acf440c4208f7ddcd6e5d819cb472e4da
Flask-WTF==1.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mongoengine==0.29.1
pymongo==4.13.2
python-dotenv==1.1.1
Werkzeug==3.1.3
WTForms==3.2.1
